requirements.txt
django==5.2.10
pillow==12.1.0
djangorestframework==3.16.1
python-decouple==3.8
drf-yasg==1.21.11
